feat: Home banner for mobile and bottom banner
