+++ assets/lang/en/banners.json
@@ -88,6 +88,6 @@
     "title": "Max out your Google privacy settings with our What Google Knows page"
   },
   "TopBarBanner": {
-    "title": "Black Friday is here! Save a huge 91% OFF all annual plans!"
+    "title": "Celebrate the holiday season with 50% OFF all lifetime plans!"
   }
 }
