+++ components/home/HeroSection.tsx
@@ -4,7 +4,7 @@ import { goToSignUpURL } from '../../lib/auth';
 import SignUpInline from '../auth/SignUpInline';
 import Image from 'next/image';
 import Header from '../shared/Header';
-import BFMainPageBanner from '../banners/BFMainPageBanner';
+import HomePageBanner from '../banners/HomePageBanner';
 
 export default function HeroSection({ textContent, lang }) {
   const [formError, setFormError] = useState<string | null>(null);
@@ -57,7 +57,7 @@ export default function HeroSection({ textContent, lang }) {
         />
         <div className="relative mx-auto flex w-full max-w-screen-xl flex-col items-center justify-between sm:mb-6 md:flex-row md:items-stretch">
           {/* Mobile view */}
-          <BFMainPageBanner />
+          <HomePageBanner />
           <div className="my-6 flex w-screen flex-shrink-0 flex-col items-center px-5 text-center sm:w-auto sm:px-0 md:my-20 md:ml-2 md:max-w-md md:items-start md:text-left lg:my-28 lg:ml-0 lg:max-w-xl">
             <Header>
               {textContent.title.line1}{' '}
@@ -83,147 +83,9 @@ export default function HeroSection({ textContent, lang }) {
           </div>
 
           {/* Desktop view full */}
-          <div className="relative ml-10 hidden max-w-2xl flex-grow md:flex xl:ml-20">
-            {/* <Image
-              loading="eager"
-              className="hidden object-contain xl:flex"
-              src="/images/home/devicesAsc.webp"
-              draggable="false"
-              width={600}
-              layout="responsive"
-              height={450}
-              alt="desktop, laptop and phone with Internxt app"
-            /> */}
 
-            {/* Browser window */}
-            <div className={`${headerAnimation1 ? 'animate-fade-in-r-large' : 'hidden'} absolute inset-0`}>
-              <div
-                className="relative left-48 top-1/2 -translate-y-1/2"
-                style={{
-                  boxShadow: '56px 56px 80px rgba(0,0,0,.2)',
-                  width: '811px',
-                  height: '490px',
-                  borderRadius: '24px',
-                }}
-              >
-                <Image
-                  loading="eager"
-                  src="/images/home/header/browser.webp"
-                  draggable="false"
-                  width={811}
-                  layout="responsive"
-                  height={490}
-                  alt="desktop, laptop and phone with Internxt app"
-                />
-              </div>
-            </div>
-
-            {/* Icon / Folder */}
-            <div
-              className={`${headerAnimation4 ? 'animate-fade-in-r' : 'hidden'} absolute left-0 top-12`}
-              style={{ filter: 'drop-shadow(8px 16px 16px rgba(0,0,0,.1))', width: '64px', height: '64px' }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/folder.svg"
-                draggable="false"
-                width={64}
-                layout="responsive"
-                height={64}
-                alt="folder icon"
-              />
-            </div>
-
-            {/* Icon / Zip */}
-            <div
-              className={`${headerAnimation5 ? 'animate-fade-in-r' : 'hidden'} absolute left-20 top-32`}
-              style={{ filter: 'drop-shadow(8px 16px 16px rgba(0,0,0,.1))', width: '56px', height: '56px' }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/zip.svg"
-                draggable="false"
-                width={56}
-                layout="responsive"
-                height={56}
-                alt="folder icon"
-              />
-            </div>
-
-            {/* Icon / Powerpoint */}
-            <div
-              className={`${headerAnimation6 ? 'animate-fade-in-r' : 'hidden'} absolute left-7 top-52`}
-              style={{ filter: 'drop-shadow(8px 16px 16px rgba(0,0,0,.1))', width: '52px', height: '52px' }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/powerpoint.svg"
-                draggable="false"
-                width={52}
-                layout="responsive"
-                height={52}
-                alt="folder icon"
-              />
-            </div>
-
-            {/* Icon / Csv */}
-            <div
-              className={`${headerAnimation7 ? 'animate-fade-in-r' : 'hidden'} absolute left-14 bottom-20`}
-              style={{ filter: 'drop-shadow(8px 16px 16px rgba(0,0,0,.1))', width: '48px', height: '48px' }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/csv.svg"
-                draggable="false"
-                width={48}
-                layout="responsive"
-                height={48}
-                alt="folder icon"
-              />
-            </div>
-
-            {/* File preview */}
-            <div
-              className={`${headerAnimation2 ? 'animate-fade-in-r-large' : 'hidden'} absolute left-20`}
-              style={{
-                boxShadow: '16px 32px 40px rgba(0,0,0,.1)',
-                width: '164px',
-                height: '178px',
-                bottom: '120px',
-                borderRadius: '24px',
-              }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/preview.webp"
-                draggable="false"
-                width={164}
-                layout="responsive"
-                height={178}
-                alt="photo file preview"
-              />
-            </div>
-
-            {/* Tasklogger */}
-            <div
-              className={`${headerAnimation3 ? 'animate-fade-in-r-large' : 'hidden'} absolute left-[416px] bottom-0`}
-              style={{
-                boxShadow: '16px 32px 40px rgba(0,0,0,.1)',
-                width: '211px',
-                height: '190px',
-                borderRadius: '24px',
-              }}
-            >
-              <Image
-                loading="eager"
-                src="/images/home/header/tasklogger.webp"
-                draggable="false"
-                width={211}
-                layout="responsive"
-                height={190}
-                alt="file logger with items downloading"
-              />
-            </div>
+          <div className="absolute right-0 hidden h-full w-full translate-x-1/2 lg:flex 2xl:translate-x-1/2">
+            <Image src="/images/lifetime/Internxt_Lifetime.webp" width={2684} height={1398} quality={100} />
           </div>
 
           {/* Desktop view reduced */}
@@ -243,7 +105,7 @@ export default function HeroSection({ textContent, lang }) {
 
         <div className="relative left-1/2 z-10 w-screen -translate-x-1/2 bg-transparent">
           <div className={'flex xl:hidden'}>
-            <Marquee gradientColor={[255, 255, 255]} gradientWidth="32px" speed={30}>
+            <Marquee gradientColor={[255, 255, 255]} className="bg-transparent" gradientWidth="32px" speed={30}>
               <div className="featured flex w-full flex-row p-6">
                 {lang === 'es' ? (
                   <a
